had issue with enemy spawn position cause by NavAgent so I only turn it on when play gets close. Cause rat swarms and game over very quickly

+++ LostAndFound2021/Assets/Character/Script/PlayerController.cs
@@ -15,7 +15,7 @@ public class PlayerController : MonoBehaviour
     public Health characterHealth;
 
     private PlayerInput inputActions;
-    public int viewDistance;
+    public float viewDistance;
     public LayerMask activationZonesMask;
 
     public Interactable currentlyInteractingWith;
